chore: update go.mod and examples
