+++ _examples/go.mod
@@ -4,4 +4,4 @@ go 1.16
 
 replace github.com/podhmo/apikit => ../
 
-require github.com/podhmo/apikit v0.0.0-00010101000000-000000000000
+require github.com/podhmo/apikit v0.0.0-20210923132720-b5e305e3689c // indirect
